Safari-611.1.10/LayoutTests/crypto/subtle/hmac-import-raw-key-customized-length.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
<script src="../resources/common.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>

<script>
description("Test importing a raw HMAC key with customized length");

jsTestIsAsync = true;

var rawKey = asciiToUint8Array("jnOw99oOZFLIEPMr");
var extractable = true;

debug("Importing a key...");
crypto.subtle.importKey("raw", rawKey, {name: "hmac", hash: "sha-1", length: 128}, extractable, ["sign", "verify"]).then(function(result) {
    key = result;

    shouldBe("key.type", "'secret'");
    shouldBe("key.extractable", "true");
    shouldBe("key.algorithm.name", "'HMAC'");
    shouldBe("key.algorithm.length", "128");
    shouldBe("key.algorithm.hash.name", "'SHA-1'");
    shouldBe("key.usages", "['sign', 'verify']");

    finishJSTest();
});

</script>

<script src="../../resources/js-test-post.js"></script>
</body>
</html>
